add delete confirm

+++ js/options.js
@@ -56,9 +56,24 @@ var app = new Vue({
 			})
         },
         clearAll(){
-			chrome.storage.sync.clear()
-            this.getSaveData()
-			console.log("已清空所有保存数据")
+            this.$confirm('此操作将永久删除所有记录，是否继续？', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                chrome.storage.sync.clear()
+                this.getSaveData()
+                console.log("已清空所有保存数据")
+                this.$message({
+                    type: 'success',
+                    message: '删除成功!'
+                });
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });          
+            });
         },
     
     }
